[Spark] Move column conversion tests into dedicated suite
